fix: use int instead of long

+++ src/training/trainer.py
@@ -268,8 +268,8 @@ class Trainer:
                     preds = preds.float()
                     labels = labels.int()
                 elif self.model_type == "slmuse-dlf":
-                    preds = preds.argmax(dim=1)
-                    labels = labels.argmax(dim=1)
+                    preds = preds.argmax(dim=1).int()
+                    labels = labels.argmax(dim=1).int()
 
                 metrics[metric][metrics_name].add_batch(
                     predictions=preds, references=labels
@@ -369,7 +369,7 @@ class Trainer:
                     logger.debug(
                         "Labels are one-hot encoded, converting to class index."
                     )
-                    prepared_labels = torch.argmax(labels, dim=1).long()
+                    prepared_labels = torch.argmax(labels, dim=1).int()
 
             if self.training_management == "accelerate":
                 with self.accelerator.accumulate(self.model):
